move DashboardRestApi & WidgetRestApi classes to internal package
